cambio de decimal a int de los pecios del festival

+++ Festival.cs
@@ -17,8 +17,8 @@ namespace ProyectoIPo
         public string Nombre { get; set; }
         public DateTime Fecha { get; set; }
         public string Ubicacion { get; set; }
-        public decimal PrecioEstandar { get; set; }
-        public decimal PrecioVIP { get; set; }
+        public int PrecioEstandar { get; set; }
+        public int PrecioVIP { get; set; }
 
         // Lista de artistas del festival, ahora contiene objetos Artista
         public ObservableCollection<Artista> Artistas { get; set; } = new ObservableCollection<Artista>();
